Классы чтения multipart-запросов перемещены в oscript.exe, т.к. используются только в CGI-режиме.
