--- src/ScriptEngine.HostedScript/Library/Http/Multipart/MultipartParseException.cs
@@ -1,51 +1,0 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="MultipartParseException.cs" company="Jake Woods">
-//   Copyright (c) 2013 Jake Woods
-//   
-//   Permission is hereby granted, free of charge, to any person obtaining a copy of this software 
-//   and associated documentation files (the "Software"), to deal in the Software without restriction, 
-//   including without limitation the rights to use, copy, modify, merge, publish, distribute, 
-//   sublicense, and/or sell copies of the Software, and to permit persons to whom the Software 
-//   is furnished to do so, subject to the following conditions:
-//   
-//   The above copyright notice and this permission notice shall be included in all copies 
-//   or substantial portions of the Software.
-//   
-//   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, 
-//   INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR 
-//   PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR 
-//   ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, 
-//   ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-// </copyright>
-// <author>Jake Woods</author>
-// <summary>
-//   Represents a parsing problem occurring within the MultipartFormDataParser
-// </summary>
-// --------------------------------------------------------------------------------------------------------------------
-
-using System;
-
-namespace HttpMultipartParser
-{
-    /// <summary>
-    ///     Represents a parsing problem occurring within the MultipartFormDataParser
-    /// </summary>
-    [Serializable]
-    internal class MultipartParseException : Exception
-    {
-        #region Constructors and Destructors
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="MultipartParseException" /> class.
-        /// </summary>
-        /// <param name="message">
-        ///     The message.
-        /// </param>
-        public MultipartParseException(string message)
-            : base(message)
-        {
-        }
-
-        #endregion
-    }
-}
