--- src/ScriptEngine.HostedScript/Library/Http/Multipart/ParameterPart.cs
@@ -1,72 +1,0 @@
-﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="ParameterPart.cs" company="Jake Woods">
-//   Copyright (c) 2013 Jake Woods
-//   
-//   Permission is hereby granted, free of charge, to any person obtaining a copy of this software 
-//   and associated documentation files (the "Software"), to deal in the Software without restriction, 
-//   including without limitation the rights to use, copy, modify, merge, publish, distribute, 
-//   sublicense, and/or sell copies of the Software, and to permit persons to whom the Software 
-//   is furnished to do so, subject to the following conditions:
-//   
-//   The above copyright notice and this permission notice shall be included in all copies 
-//   or substantial portions of the Software.
-//   
-//   THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, 
-//   INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR 
-//   PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR 
-//   ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, 
-//   ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-// </copyright>
-// <author>Jake Woods</author>
-// <summary>
-//   Represents a single parameter extracted from a multipart/form-data
-//   stream.
-// </summary>
-// --------------------------------------------------------------------------------------------------------------------
-
-namespace HttpMultipartParser
-{
-    /// <summary>
-    ///     Represents a single parameter extracted from a multipart/form-data
-    ///     stream.
-    /// </summary>
-    /// <remarks>
-    ///     For our purposes a "parameter" is defined as any non-file data
-    ///     in the multipart/form-data stream.
-    /// </remarks>
-    public class ParameterPart
-    {
-        #region Constructors and Destructors
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="ParameterPart" /> class.
-        /// </summary>
-        /// <param name="name">
-        ///     The name.
-        /// </param>
-        /// <param name="data">
-        ///     The data.
-        /// </param>
-        public ParameterPart(string name, string data)
-        {
-            Name = name;
-            Data = data;
-        }
-
-        #endregion
-
-        #region Public Properties
-
-        /// <summary>
-        ///     Gets the data.
-        /// </summary>
-        public string Data { get; private set; }
-
-        /// <summary>
-        ///     Gets or sets the name.
-        /// </summary>
-        public string Name { get; set; }
-
-        #endregion
-    }
-}
