+++ src/ScriptEngine.HostedScript/ScriptEngine.HostedScript.csproj
@@ -98,15 +98,6 @@
     <Compile Include="Library\Http\HttpResponseBody.cs" />
     <Compile Include="Library\Http\HttpResponseContext.cs" />
     <Compile Include="Library\Http\InternetProxyContext.cs" />
-    <Compile Include="Library\Http\Multipart\BinaryStreamStack.cs" />
-    <Compile Include="Library\Http\Multipart\FilePart.cs" />
-    <Compile Include="Library\Http\Multipart\MultipartFormDataParser.cs" />
-    <Compile Include="Library\Http\Multipart\MultipartParseException.cs" />
-    <Compile Include="Library\Http\Multipart\ParameterPart.cs" />
-    <Compile Include="Library\Http\Multipart\PostFileDescription.cs" />
-    <Compile Include="Library\Http\Multipart\RebufferableBinaryReader.cs" />
-    <Compile Include="Library\Http\Multipart\StreamingMultipartFormDataParser.cs" />
-    <Compile Include="Library\Http\Multipart\SubsequenceFinder.cs" />
     <Compile Include="Library\ConsoleColorEnum.cs" />
     <Compile Include="Library\MessageStatusEnum.cs" />
     <Compile Include="Library\MiscGlobalFunctions.cs" />
@@ -163,7 +154,6 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Library\StructureImpl.cs" />
     <Compile Include="Library\StringOperations.cs" />
-    <Compile Include="Library\Http\Multipart\PostRequestData.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\ScriptEngine\ScriptEngine.csproj">
