+++ src/oscript/CgiBehavior.cs
@@ -58,7 +58,7 @@ namespace oscript
             engine.CustomConfig = ScriptFileHelper.CustomConfigPath(scriptFile);
             engine.AttachAssembly(System.Reflection.Assembly.GetExecutingAssembly());
 
-            var request = new WebRequestContext();
+            var request = new Web.WebRequestContext();
             engine.InjectGlobalProperty("ВебЗапрос", request, true);
             engine.InjectGlobalProperty("WebRequest", request, true);
             engine.InjectObject(this, false);
