+++ src/oscript/Web/WebRequestContext.cs
@@ -4,19 +4,15 @@ Mozilla Public License, v.2.0. If a copy of the MPL
 was not distributed with this file, You can obtain one 
 at http://mozilla.org/MPL/2.0/.
 ----------------------------------------------------------*/
-using ScriptEngine.HostedScript.Library;
-using ScriptEngine.Machine;
-using ScriptEngine.Machine.Contexts;
 using System;
 using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text;
-using System.IO;
-using HttpMultipartParser;
-using ScriptEngine.HostedScript.Library.Http.Multipart;
+using ScriptEngine.Machine;
+using ScriptEngine.Machine.Contexts;
+using ScriptEngine.HostedScript.Library;
+using oscript.Web.Multipart;
 
-namespace oscript
+namespace oscript.Web
 {
     [ContextClass ("ВебЗапрос", "WebRequest")]
     public class WebRequestContext : AutoContext<WebRequestContext>
